Move abstract Client to network module (io.luna.net.client -> io.luna.net)
This also reverts the updated network package name to avoid a huge commit of import changes
