+++ network/src/main/java/io/luna/net/codec/ByteOrder.java
@@ -1,4 +1,4 @@
-package io.luna.network.codec;
+package io.luna.net.codec;
 
 /**
  * An enum representing endianness of game messages.
